server-endpoints/requirements.txt
# Dependencias para endpoints Stripe en CaloriasAPI
fastapi>=0.104.0
uvicorn>=0.24.0
stripe>=7.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
supabase>=2.0.0